// src/models/IoK8sApiFlowcontrolV1beta3PriorityLevelConfigurationList/types.ts
import { IoK8sApiFlowcontrolV1beta3PriorityLevelConfigurationModelOptions as IoK8sApiFlowcontrolV1beta3PriorityLevelConfiguration } from '../IoK8sApiFlowcontrolV1beta3PriorityLevelConfiguration';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiFlowcontrolV1beta3PriorityLevelConfigurationListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiFlowcontrolV1beta3PriorityLevelConfiguration>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}